+++ src/extension.ts
@@ -26,6 +26,7 @@ import { registerToGqlFragmentToDart } from './vscode_explore_menu/graphql_fragm
 import { registerCreateRouteConfiguration } from './vscode_explorer/flutter/generate_route_temp';
 import { registerCleanArchitectureCubitGenerate } from './vscode_explorer/flutter/generate_clean_architecture_cubit';
 import { registerFlutterPageGenerate } from './vscode_explorer/flutter/generate_flutter_page';
+import { runCommand } from './utils/src/terminal_utils/terminal_utils';
 let sidebarManger = new SidebarManager()
 export class APP {
   public static pubspecYaml: any|undefined = undefined;
@@ -34,8 +35,8 @@ export class APP {
   public static depOhive: any|undefined = undefined;
   public static flutterLibName : any|undefined = undefined;
   public static flutterLocalizations : any|undefined = undefined;
+  public static myName = "";
   public static depOnLogging : any|undefined = undefined;
-
 
 }
 
@@ -46,13 +47,14 @@ export async function activate(context: vscode.ExtensionContext) {
   updateGitSubModule(context)
   await checkGitExtensionInYamlIfDart(true).then(async (yaml) => {
     APP.pubspecYaml = yaml
+    APP.myName = await runCommand("whoami")
+    APP.myName = APP.myName.replace("\n","")
     if(yaml!=undefined){
       APP.pubspecLockYaml =await getPubspecLockAsMap()
       APP.depOnBloc = APP.pubspecYaml["dependencies"]["flutter_bloc"] !=undefined
       APP.depOhive = APP.pubspecYaml["dependencies"]["hive"] !=undefined
       APP.flutterLibName = APP.pubspecYaml["name"]
       APP.flutterLocalizations = APP.pubspecYaml["dependencies"]["flutter_localizations"]
-      APP.depOnLogging = APP.pubspecYaml["dependencies"]["logging"]
       APP.depOnLogging = APP.pubspecYaml["dependencies"]["color_logging"]
   
       log(APP.pubspecYaml)